This fixes a serious bug that exists in the upstream code currently and is quite hard to spot:
The `input` and `indexMatrix` follow a different indexing convention than the output matrix, even though
all of these structures share the same type (vectors of 2D Matrices).
Concretely, the vector index in `input` and `indexMatrix` corresponds to the event in the batch. The number of
rows in each of their matrices corresponds to the depth, and the number of columns corresponds to the localViews,
i.e to the product of `width * height`. The output however has no notion of a batch. The vector indexing corresponds
to the depth, and the rows and columns of each matrix correspond to the actual rows and columns of the output.

However the original code (line 179 at `MaxPoolLayer.h`) seems to assume that the output is also indexed by the batch
number which leads to a serious and hard to spot bug (the current tests cannot catch it since the number of batches
is equal to the depth and equal to 1). The current commit fixes this misunderstanding (along with other issues in the tests)